+++ client/src/components/favorite-card/FavoriteCard.css
